# seat-in-heaven-express/openapi.json
{"components":{"schemas":{"OkResponse":{"properties":{"ok":{"type":"boolean"}},"type":"object","required":["ok"]},"PostAuthBody":{"properties":{"email":{"maxLength":255,"type":"string","format":"email","minLength":1},"password":{"maxLength":255,"type":"string","minLength":1}},"type":"object","required":["email","password"]},"FloorBody":{"properties":{"floorname":{"maxLength":50,"type":"string","minLength":1}},"type":"object","required":["floorname"]},"PostFloorBody":{"properties":{"floortype":{"type":"string","enum":["FLOOR","ROOM"],"minLength":1},"floorname":{"maxLength":50,"type":"string","minLength":1},"parent_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"}},"type":"object","required":["floortype","floorname","parent_id"]},"FloorsPathParams":{"properties":{"floor_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"}},"type":"object","required":["floor_id"]},"FloorsQuery":{"properties":{"floor_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"}},"type":"object"},"FloorsOrderBody":{"properties":{"floor_id_list":{"items":{"exclusiveMinimum":true,"minimum":0,"type":"number"},"type":"array"}},"type":"object","required":["floor_id_list"]},"FloorsNodeBody":{"properties":{"parent_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"},"floor_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"}},"type":"object","required":["parent_id","floor_id"]},"FloorResponse":{"properties":{"floor_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"},"floortype":{"type":"string","enum":["FLOOR","ROOM"],"minLength":1},"floorname":{"maxLength":50,"type":"string","minLength":1},"order":{"type":"integer"},"updated_at":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"}},"type":"object","required":["floor_id","floortype","floorname","order","updated_at"]},"FloorResponseWithChildren":{"properties":{"children":{"items":{"$ref":"#/components/schemas/FloorResponse"},"type":"array"},"floor_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"},"floortype":{"type":"string","enum":["FLOOR","ROOM"],"minLength":1},"floorname":{"maxLength":50,"type":"string","minLength":1},"order":{"type":"integer"},"updated_at":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"}},"type":"object","required":["children","floor_id","floortype","floorname","order","updated_at"]},"RootFloorResponse":{"properties":{"ancestors":{"items":{"$ref":"#/components/schemas/FloorResponse"},"type":"array"},"children":{"items":{"$ref":"#/components/schemas/FloorResponseWithChildren"},"type":"array"},"floor_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"},"floortype":{"type":"string","enum":["FLOOR","ROOM"],"minLength":1},"floorname":{"maxLength":50,"type":"string","minLength":1},"order":{"type":"integer"},"updated_at":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"}},"type":"object","required":["ancestors","children","floor_id","floortype","floorname","order","updated_at"]},"RoomsQuery":{"properties":{"keyword":{"maxLength":255,"type":"string"}},"type":"object"},"RoomFloorResponse":{"properties":{"ancestors":{"items":{"$ref":"#/components/schemas/FloorResponse"},"type":"array"},"floor_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"},"floortype":{"type":"string","enum":["FLOOR","ROOM"],"minLength":1},"floorname":{"maxLength":50,"type":"string","minLength":1},"order":{"type":"integer"},"updated_at":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"}},"type":"object","required":["ancestors","floor_id","floortype","floorname","order","updated_at"]},"ListRoomFloorResponse":{"properties":{"rooms":{"items":{"$ref":"#/components/schemas/RoomFloorResponse"},"type":"array"}},"type":"object","required":["rooms"]},"UserBody":{"properties":{"email":{"format":"email","type":"string","maxLength":255,"minLength":1},"username":{"maxLength":50,"type":"string","minLength":1},"role":{"type":"string","enum":["SYSTEM_ADMIN","LIMITED_ADMIN","USER"],"minLength":1}},"type":"object","required":["email","username","role"]},"UserPathParams":{"properties":{"user_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"}},"type":"object","required":["user_id"]},"UsersQuery":{"properties":{"keyword":{"maxLength":255,"type":"string"},"roles":{"items":{"type":"string","enum":["SYSTEM_ADMIN","LIMITED_ADMIN","USER"]},"type":"array"},"limit":{"type":"string","minLength":1},"offset":{"type":"string","minLength":1}},"type":"object","required":["limit","offset"]},"UserResponse":{"properties":{"user_id":{"exclusiveMinimum":true,"minimum":0,"type":"number"},"email":{"format":"email","type":"string","maxLength":255,"minLength":1},"username":{"maxLength":50,"type":"string","minLength":1},"role":{"type":"string","enum":["SYSTEM_ADMIN","LIMITED_ADMIN","USER"],"minLength":1},"updated_at":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"}},"type":"object","required":["user_id","email","username","role","updated_at"]},"ListUserResponse":{"properties":{"total":{"type":"integer"},"users":{"items":{"$ref":"#/components/schemas/UserResponse"},"type":"array"}},"type":"object","required":["total","users"]}}},"info":{"title":"","version":"1.0.0"},"openapi":"3.0.0","paths":{"/api/auth":{"get":{"operationId":"AuthController.getAuth","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponse"}}},"description":""}},"summary":"Get auth","tags":["Auth"]},"post":{"operationId":"AuthController.postAuth","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostAuthBody"}}},"description":"PostAuthBody","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponse"}}},"description":""}},"summary":"Post auth","tags":["Auth"]},"delete":{"operationId":"AuthController.deleteAuth","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponse"}}},"description":""}},"summary":"Delete auth","tags":["Auth"]}},"/api/floors":{"post":{"operationId":"FloorsController.postFloor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostFloorBody"}}},"description":"PostFloorBody","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorResponse"}}},"description":""}},"summary":"Post floor","tags":["Floors"]},"get":{"operationId":"FloorsController.getFloors","parameters":[{"in":"query","name":"floor_id","schema":{"exclusiveMinimum":true,"minimum":0,"type":"number"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RootFloorResponse"}}},"description":""}},"summary":"Get floors","tags":["Floors"]}},"/api/floors/{floor_id}":{"put":{"operationId":"FloorsController.putFloor","parameters":[{"in":"path","name":"floor_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/FloorBody"},{"properties":{"updated_at":{"type":"string"}},"required":["updated_at"],"type":"object"}]}}},"description":"FloorBody","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorResponse"}}},"description":""}},"summary":"Put floor","tags":["Floors"]},"get":{"operationId":"FloorsController.getFloor","parameters":[{"in":"path","name":"floor_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorResponse"}}},"description":""}},"summary":"Get floor","tags":["Floors"]},"delete":{"operationId":"FloorsController.deleteFloor","parameters":[{"in":"path","name":"floor_id","required":true,"schema":{"type":"string"}},{"in":"query","name":"updated_at","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponse"}}},"description":""}},"summary":"Delete floor","tags":["Floors"]}},"/api/floors/order":{"patch":{"operationId":"FloorsController.patchFloorsOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorsOrderBody"}}},"description":"FloorsOrderBody","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponse"}}},"description":""}},"summary":"Patch floors order","tags":["Floors"]}},"/api/floors/node":{"patch":{"operationId":"FloorsController.patchFloorsNode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorsNodeBody"}}},"description":"FloorsNodeBody","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponse"}}},"description":""}},"summary":"Patch floors node","tags":["Floors"]}},"/api/rooms":{"post":{"operationId":"RoomsController.postFloor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostFloorBody"}}},"description":"PostFloorBody","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorResponse"}}},"description":""}},"summary":"Post floor","tags":["Rooms"]},"get":{"operationId":"RoomsController.getRooms","parameters":[{"in":"query","name":"keyword","schema":{"maxLength":255,"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListRoomFloorResponse"}}},"description":""}},"summary":"Get rooms","tags":["Rooms"]}},"/api/rooms/{room_id}":{"put":{"operationId":"RoomsController.putFloor","parameters":[{"in":"path","name":"room_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/FloorBody"},{"properties":{"updated_at":{"type":"string"}},"required":["updated_at"],"type":"object"}]}}},"description":"FloorBody","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorResponse"}}},"description":""}},"summary":"Put floor","tags":["Rooms"]},"get":{"operationId":"RoomsController.getFloor","parameters":[{"in":"path","name":"room_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorResponse"}}},"description":""}},"summary":"Get floor","tags":["Rooms"]},"delete":{"operationId":"RoomsController.deleteFloor","parameters":[{"in":"path","name":"room_id","required":true,"schema":{"type":"string"}},{"in":"query","name":"updated_at","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OkResponse"}}},"description":""}},"summary":"Delete floor","tags":["Rooms"]}},"/api/users":{"post":{"operationId":"UsersController.postUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBody"}}},"description":"UserBody","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}},"description":""}},"summary":"Post user","tags":["Users"]},"get":{"operationId":"UsersController.getUsers","parameters":[{"in":"query","name":"keyword","required":false,"schema":{"maxLength":255,"type":"string"}},{"in":"query","name":"roles","required":false,"schema":{"items":{"type":"string","enum":["SYSTEM_ADMIN","LIMITED_ADMIN","USER"]},"type":"array"}},{"in":"query","name":"limit","required":true,"schema":{"type":"string","minLength":1}},{"in":"query","name":"offset","required":true,"schema":{"type":"string","minLength":1}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListUserResponse"}}},"description":""}},"summary":"Get users","tags":["Users"]}},"/api/users/{user_id}":{"put":{"operationId":"UsersController.putUser","parameters":[{"in":"path","name":"user_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/UserBody"},{"properties":{"updated_at":{"type":"string"}},"required":["updated_at"],"type":"object"}]}}},"description":"UserBody","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}},"description":""}},"summary":"Put user","tags":["Users"]},"get":{"operationId":"UsersController.getUser","parameters":[{"in":"path","name":"user_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}},"description":""}},"summary":"Get user","tags":["Users"]},"delete":{"operationId":"UsersController.deleteUser","parameters":[{"in":"path","name":"user_id","required":true,"schema":{"type":"string"}},{"in":"query","name":"updated_at","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}},"description":""}},"summary":"Delete user","tags":["Users"]}},"/openapi":{"get":{"operationId":"OpenapiController.getOpenApi","responses":{"200":{"content":{"application/json":{}},"description":"Successful response"}},"summary":"Get open api","tags":["Openapi"]}}}}